Fix wrong data type of the direction property when read from file

+++ src/napari_bee_annotator/_reader.py
@@ -74,7 +74,7 @@ def reader_function(path):
     for directions, data in zip(direction_arrays, data_arrays):
         # optional kwargs for the corresponding viewer.add_* method
         add_kwargs = {
-            "properties": {"Direction": directions},
+            "properties": {"Direction": directions.astype(np.uint32)},
             "colormap": "viridis",
             "color_by": "Direction",
             "tail_width": 5,
